quality_to_steps_map is not only for fal-ai
