+++ pipelex/cogt/config_cogt.py
@@ -1,21 +1,45 @@
-from pydantic import Field
+from pydantic import Field, field_validator
 
 from pipelex.cogt.exceptions import LLMConfigError
-from pipelex.cogt.img_gen.img_gen_job_components import ImgGenJobConfig, ImgGenJobParams, ImgGenJobParamsDefaults
+from pipelex.cogt.img_gen.img_gen_job_components import ImgGenJobConfig, ImgGenJobParams, ImgGenJobParamsDefaults, Quality
 from pipelex.cogt.llm.llm_job_components import LLMJobConfig
 from pipelex.cogt.models.model_deck_config import ModelDeckConfig
-from pipelex.plugins.fal.fal_config import FalConfig
 from pipelex.system.configuration.config_model import ConfigModel
+from pipelex.system.exceptions import ConfigValidationError
 
 
 class ExtractConfig(ConfigModel):
     default_page_views_dpi: int
 
 
+class QualityToStepsMap(ConfigModel):
+    flux_map_quality_to_steps: dict[str, int]
+    sdxl_lightning_map_quality_to_steps: dict[str, int]
+    qwen_image_map_quality_to_steps: dict[str, int]
+
+    @field_validator("flux_map_quality_to_steps", "sdxl_lightning_map_quality_to_steps", "qwen_image_map_quality_to_steps")
+    @classmethod
+    def validate_quality_mapping(cls, value: dict[str, int]) -> dict[str, int]:
+        valid_qualities = {quality.value for quality in Quality}
+        missing_qualities = valid_qualities - set(value.keys())
+        invalid_qualities = set(value.keys()) - valid_qualities
+
+        if missing_qualities and invalid_qualities:
+            msg = f"Missing ({missing_qualities}) and invalid ({invalid_qualities}) quality levels in mapping"
+            raise ConfigValidationError(msg)
+        if missing_qualities:
+            msg = f"Missing quality levels in mapping: {missing_qualities}"
+            raise ConfigValidationError(msg)
+        if invalid_qualities:
+            msg = f"Invalid quality levels in mapping: {invalid_qualities}"
+            raise ConfigValidationError(msg)
+        return value
+
+
 class ImgGenConfig(ConfigModel):
     img_gen_job_config: ImgGenJobConfig
     img_gen_param_defaults: ImgGenJobParamsDefaults
-    fal_config: FalConfig
+    quality_to_steps_map: QualityToStepsMap
 
     def make_default_img_gen_job_params(self) -> ImgGenJobParams:
         return self.img_gen_param_defaults.make_img_gen_job_params()
